packages/schema/python/pyproject.toml
[project]
name = "swael-schema"
version = "1.0.2"
description = "SwarmKit JSON Schemas + Python validator."
readme = {text = "SwarmKit canonical JSON Schemas + Python validator. See https://github.com/delivstat/swarmkit", content-type = "text/markdown"}
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Srijith Kartha" }]
keywords = ["swarmkit", "schema", "json-schema"]

dependencies = [
    "jsonschema>=4.23",
    "pydantic>=2.8",
    "pyyaml>=6.0",
    "referencing>=0.35",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "mypy>=1.11",
    "ruff>=0.6",
    "types-pyyaml",
    "types-jsonschema",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/swael_schema"]
